Basic shape morphing (including solid-fill morphing) works now, I think.
- Create abstract base classes for fill_style and line_style to share with
  their morph variants.  Add ratio parameter (unused in non-morph classes)
  to apply().

- Move morph_{line,fill}_style declarations to gameswf_styles, though impls
  are still in _morph.

Added test_shape_tweening-2.swf, which seems to work basically as it
does in flashplayer.  The existing test_shape_tweening.swf displays a
strange artifact seemingly related to one vertex flying off in the
wrong direction, as well as what appears to be filling of the wrong
side of the shape.  It's very curious, and I haven't been able to figure it
out yet.

No caching of the meshes, so we re-tesselate on every display, irrespective
of changes (or lack thereof) in ratio or error limit.  Also no reasonable
morphing of line styles, gradients, bitmap fills (!?), etc.