New triangulator is usable in gameswf.  Seems to be fairly robust based on playing a number of animation-heavy SWFs off the web.
Still a few niggling problems to look into, but it's probably no
worse than the existing tesselator.

The interface code is currently very inefficient; needs basic
mechanical cleanup and optimization.

Also, need to update the gameswf code to use the new triangulator for
rendering font bitmaps, remove the tristripper, and figure out what's
wrong with gameswf_morph2.

Parameterized the new triangulator so it works with floats or sint16's.
Initial testing suggests that it is very fast.